v1.1.0 - refactor config module
