+++ src/config/generators/mod.rs
@@ -1,0 +1,3 @@
+mod solid;
+
+pub(super) use solid::SolidConfig;
