+++ src/config/mod.rs
@@ -1,0 +1,6 @@
+mod config;
+mod generators;
+mod user_config;
+
+pub use config::Config;
+pub use user_config::Generators;
